Clean up commit
