+++ src/modules/like.js
@@ -1,48 +1,30 @@
-import { fetchLikesFromAPI, updateLikesOnAPI } from './api.js';
+import { invAPI } from "./api.js";
 
-const invAPI = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/app/dSv5DdTGl6SZHdXDAlEr';
-const likes = {};
+const likeURL = invAPI + 'likes';
 
-const saveLikesToAPI = async (likesData) => {
-    const res = await fetch(`${invAPI}/likes`, {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(likesData),
+const addLike = async (id) => {
+    await fetch(likeURL, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        item_id: id,
+      }),
     });
-    const data = res.text();
-    return data;
-}
-
-const updateLikes = (item_id) => {
-    const likesCountElement = document.querySelector(`[data-like="${item_id}"] .likeCount`);
-    if (likesCountElement) {
-        likes[item_id] = (likes[item_id] || 0) + 1;
-        likesCountElement.textContent = `Likes: ${likes[item_id]}`;
-        updateLikesOnAPI(item_id);
+};
+  
+ const getLikes = async (id) => {
+    const res = await fetch(likeURL);
+    const data = await res.json();
+    const findItem = data.find(({item_id}) => item_id == id);
+    if (findItem) {
+      return findItem.likes;
+    } else {
+      return 0
     }
-}
+};
 
-const updateAllLikes = (item_id) => {
-    for (const item_id in likes) {
-        const likesCountElement = document.querySelector(`[data-like="${item_id}"] .likeCount`);
-        if (likesCountElement) {
-            likesCountElement.textContent = `Likes: ${likes[item_id]}`;
-        }
-    }
-}
+export {addLike , getLikes}
 
-const fetchLikesFromAPIExternal = async () => {
-    try {
-        const response = await fetch(`${invAPI}/likes`);
-        const data = await response.json();
-        Object.assign(likes, data);
-        updateAllLikes();
-    } catch (error) {
-        console.error('Error fetching likes from API:', error);
-    }
-}
-
-export { updateLikes, updateAllLikes, fetchLikesFromAPIExternal , saveLikesToAPI }
